+++ src/index.ts
@@ -73,7 +73,9 @@ export * from './components/Carousel';
 export * from './components/Detail';
 export * from './patterns/DetailList';
 export * from './components/PinchZoom';
+export * from './components/Tabs';
 export * from './components/SeldonImage';
+export * from './patterns/SaleHeaderBanner';
 
 // utils
 export * from './utils/hooks';