+++ src/patterns/SaleHeaderBanner/types.ts
@@ -1,0 +1,5 @@
+export enum AuctionState {
+  preSale = 'PRESALE',
+  openForBidding = 'OPEN_FOR_BIDDING',
+  past = 'PAST',
+}